test: 单元测试

+++ journey-mybatis-complete/src/main/java/org/example/mybatis/entity/User.java
@@ -2,24 +2,24 @@ package org.example.mybatis.entity;
 
 public class User {
 
-    private Integer id;
+    private Long id;
     private String username;
     private String password;
 
     public User() {
     }
 
-    public User(Integer id, String username, String password) {
+    public User(Long id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
